# Sprint 2.1.5 Backend Phase 1: Lead Protection & Progressive Profiling