… backend/MIGRATIONS.md
… docs/planung/CRM_COMPLETE_MASTER_PLAN_V5.md
… docs/planung/DOCUMENTATION_CHECKLIST.md
… docs/planung/PRODUCTION_ROADMAP_2025.md
… docs/planung/TRIGGER_SPRINT_2_1_5.md
… docs/planung/TRIGGER_SPRINT_2_1_6.md
… docs/planung/features-neu/02_neukundengewinnung/SPRINT_MAP.md
… docs/planung/features-neu/02_neukundengewinnung/artefakte/SPRINT_2_1_4/SUMMARY.md
… docs/planung/features-neu/02_neukundengewinnung/artefakte/SPRINT_2_1_5/CONTRACT_MAPPING.md
… docs/planung/features-neu/02_neukundengewinnung/artefakte/SPRINT_2_1_5/DELTA_LOG_2_1_5.md
… docs/planung/features-neu/02_neukundengewinnung/artefakte/SPRINT_2_1_5/SUMMARY.md
… docs/planung/features-neu/02_neukundengewinnung/artefakte/SPRINT_2_1_5/TEST_PLAN.md
… docs/planung/features-neu/02_neukundengewinnung/backend/_index.md
… docs/planung/features-neu/02_neukundengewinnung/shared/adr/ADR-004-lead-protection-inline-first.md